rag-mcp/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
qdrant-client>=1.12.0
sentence-transformers>=3.3.0
mcp>=1.0.0
sse-starlette>=2.0.0
